requirements.txt
#For ranking
numpy
pandas
streamlit
plotly
openelo

#For parser

camelot-py[base]
